# packages/shared/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat/{characterId}": {
      "post": {
        "operationId": "NarrativeController_sendMessage",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "AI response with narrative and instructions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a chat message and get AI response",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{characterId}/history": {
      "get": {
        "operationId": "NarrativeController_getNarrativeHistory",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Narrative history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Narrative not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get narrative history with a character",
        "tags": [
          "chat"
        ]
      },
      "delete": {
        "operationId": "NarrativeController_clearNarrative",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Narrative cleared"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear narrative history",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{characterId}/messages": {
      "get": {
        "operationId": "NarrativeController_getRecentMessages",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NarrativeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get recent messages from narrative",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/combat/{characterId}/start": {
      "post": {
        "operationId": "CombatController_startCombat",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CombatStartRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initialize combat with enemies",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/combat/{characterId}/action": {
      "post": {
        "operationId": "CombatController_action",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CombatActionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatActionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute any combat action (attack, dash, disengage, spell, class feature)",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/combat/{characterId}/status": {
      "get": {
        "operationId": "CombatController_getStatus",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current combat status",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/combat/{characterId}/end-turn": {
      "post": {
        "operationId": "CombatController_endTurn",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns attack logs for animations and new player turn state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndPlayerTurnResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "End current player activation and advance turn (triggers enemy actions)",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/combat/{characterId}/flee": {
      "post": {
        "operationId": "CombatController_flee",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatEndResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Force end current combat (flee)",
        "tags": [
          "combat"
        ]
      }
    },
    "/api/combat/{characterId}/move": {
      "post": {
        "operationId": "CombatController_move",
        "parameters": [
          {
            "name": "characterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovementRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute combatant movement on the grid",
        "tags": [
          "combat"
        ]
      }
    },
...
  },
  "info": {
    "title": "RPG Gen Portal",
    "description": "Prototype API for RPG portal using Gemini tokens",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChatMessageRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message text from user",
            "example": "I want to attack the goblin"
          }
        },
        "required": [
          "message"
        ]
      },
      "NarrativeResponseDto": {
        "type": "object",
        "properties": {
          "narrative": {
            "type": "string",
            "description": "Narrative narrative content"
          },
          "instructions": {
            "description": "Game instructions",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "narrative",
          "instructions"
        ]
      },
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "characterId": {
            "type": "string",
            "description": "Character ID"
          },
          "sessionId": {
            "type": "string",
            "description": "Session ID"
          },
          "messages": {
            "description": "Messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NarrativeResponseDto"
            }
          }
        },
        "required": [
          "userId",
          "characterId",
          "sessionId",
          "messages"
        ]
      },
...
      "DiceResultDto": {
        "type": "object",
        "properties": {
          "rolls": {
            "description": "Individual dice roll results",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "modifierValue": {
            "type": "number",
            "description": "Modifier applied to the total"
          },
          "total": {
            "type": "number",
            "description": "Total result (sum of rolls + modifier)"
          }
        },
        "required": [
          "rolls",
          "modifierValue",
          "total"
        ]
      },
      "CombatStartEntryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Enemy name"
          },
          "hp": {
            "type": "number",
            "description": "Enemy HP"
          },
          "ac": {
            "type": "number",
            "description": "Enemy AC"
          },
          "attack_bonus": {
            "type": "number",
            "description": "Attack bonus (optional)"
          },
          "damage_dice": {
            "type": "string",
            "description": "Damage dice (optional)"
          },
          "damage_bonus": {
            "type": "number",
            "description": "Damage bonus (optional)"
          }
        },
        "required": [
          "name",
          "hp",
          "ac"
        ]
      },
      "CombatStartRequestDto": {
        "type": "object",
        "properties": {
          "combat_start": {
            "description": "Array of enemies to initialize combat with",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatStartEntryDto"
            }
          }
        },
        "required": [
          "combat_start"
        ]
      },
...
      "CombatantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the combatant (player character or enemy)"
          },
          "name": {
            "type": "string",
            "description": "Combatant name"
          },
          "initiative": {
            "type": "number",
            "description": "Initiative order value (finesse-based)"
          },
          "isPlayer": {
            "type": "boolean",
            "description": "Whether combatant is player character"
          },
          "hp": {
            "type": "number",
            "description": "Current hit points"
          },
          "hpMax": {
            "type": "number",
            "description": "Maximum hit points"
          },
          "pa": {
            "type": "number",
            "description": "Current action points (PA)"
          },
          "paMax": {
            "type": "number",
            "description": "Maximum action points (PA)"
          },
          "pm": {
            "type": "number",
            "description": "Current movement points (PM)"
          },
          "pmMax": {
            "type": "number",
            "description": "Maximum movement points (PM)"
          },
          "level": {
            "type": "number",
            "description": "Combatant level (1-20)"
          },
          "className": {
            "type": "string",
            "description": "Class name (guerrier, rogue, mage)"
          },
          "basePower": {
            "type": "number",
            "description": "Base power for attacks (used in damage formula)"
          },
          "scalingAttribute": {
            "type": "string",
            "description": "Which attribute scales damage (vigor, finesse, mind, survival)"
          },
          "stats": {
            "description": "Combat stats (vigor, finesse, mind, survival)",
            "allOf": [
              {
                "$ref": "#/components/schemas/TacticalStats"
              }
            ]
          },
          "side": {
            "type": "string",
            "description": "Combat side (player or enemy)"
          },
          "position": {
            "type": "object",
            "description": "Grid position for tactical combat"
          }
        },
        "required": [
          "id",
          "name",
          "initiative",
          "isPlayer"
        ]
      },
      "CombatEndDto": {
        "type": "object",
        "properties": {
          "victory": {
            "type": "boolean",
            "description": "Victory state"
          },
          "xp_gained": {
            "type": "number",
            "description": "XP gained"
          },
          "player_hp": {
            "type": "number",
            "description": "Player's HP at the end"
          },
          "enemies_defeated": {
            "description": "Enemies that were defeated",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fled": {
            "type": "boolean",
            "description": "Flee indicator"
          }
        },
        "required": [
          "victory",
          "xp_gained",
          "player_hp",
          "enemies_defeated"
        ]
      },
      "CombatStateDto": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "description": "Character ID"
          },
          "inCombat": {
            "type": "boolean",
            "description": "Whether currently in combat"
          },
          "enemies": {
            "description": "Active enemies",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatantDto"
            }
          },
          "player": {
            "description": "Player state",
            "allOf": [
              {
                "$ref": "#/components/schemas/CombatantDto"
              }
            ]
          },
          "turnOrder": {
            "description": "Turn order for combat",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatantDto"
            }
          },
          "currentTurnIndex": {
            "type": "number",
            "description": "Index of current turn in turnOrder"
          },
          "roundNumber": {
            "type": "number",
            "description": "Current round number"
          },
          "narrative": {
            "type": "string",
            "description": "Narrative summary of current combat"
          },
          "activeEffects": {
            "description": "Active status effects (stunned, burning, etc.)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "combatEnd": {
            "description": "Combat end result, populated when combat ends (inCombat=false)",
            "allOf": [
              {
                "$ref": "#/components/schemas/CombatEndDto"
              }
            ]
          }
        },
        "required": [
          "characterId",
          "inCombat",
          "enemies",
          "player",
          "turnOrder",
          "currentTurnIndex",
          "roundNumber"
        ]
      },
      "CombatActionRequestDto": {
        "type": "object",
        "properties": {
          "aptitudeId": {
            "type": "string",
            "description": "Aptitude ID to use (including basic attack, dash, etc.)"
          },
          "targetId": {
            "type": "string",
            "description": "Target combatant ID (for aptitudes targeting enemies)"
          }
        },
        "required": [
          "aptitudeId"
        ]
      },
      "CombatDiceResultDto": {
        "type": "object",
        "properties": {
          "rolls": {
            "description": "Individual dice roll results",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "modifierValue": {
            "type": "number",
            "description": "Modifier applied to the total"
          },
          "total": {
            "type": "number",
            "description": "Total result (sum of rolls + modifier)"
          },
          "damageTotal": {
            "type": "number",
            "description": "Total damage dealt"
          },
          "isCrit": {
            "type": "boolean",
            "description": "Whether this was a critical hit"
          }
        },
        "required": [
          "rolls",
          "modifierValue",
          "total",
          "damageTotal",
          "isCrit"
        ]
      },
      "CombatActionResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the action was successful"
          },
          "cost": {
            "type": "string",
            "description": "Cost of the action",
            "enum": [
              "action",
              "bonus-action",
              "reaction",
              "free"
            ]
          },
          "hit": {
            "type": "boolean",
            "description": "Whether attack/spell hit (if applicable)"
          },
          "damage": {
            "type": "number",
            "description": "Damage dealt (if applicable)"
          },
          "healing": {
            "type": "number",
            "description": "Healing restored (if applicable)"
          },
          "description": {
            "type": "string",
            "description": "Description of action result"
          },
          "errorMessage": {
            "type": "string",
            "description": "Error message if action failed"
          },
          "diceResult": {
            "description": "Dice roll result (for attacks)",
            "allOf": [
              {
                "$ref": "#/components/schemas/DiceResultDto"
              }
            ]
          },
          "damageDiceResult": {
            "description": "Damage dice result details",
            "allOf": [
              {
                "$ref": "#/components/schemas/CombatDiceResultDto"
              }
            ]
          },
          "damageTotal": {
            "type": "number",
            "description": "Total damage dealt (convenience field)"
          },
          "isCrit": {
            "type": "boolean",
            "description": "Whether the attack was a critical hit"
          },
          "narrative": {
            "type": "string",
            "description": "Narrative text (e.g., for combat end)"
          }
        },
        "required": [
          "success",
          "cost"
        ]
      },
      "EnemyAttackLogDto": {
        "type": "object",
        "properties": {
          "attackerId": {
            "type": "string",
            "description": "ID of the attacking enemy",
            "example": "goblin-1"
          },
          "attackerName": {
            "type": "string",
            "description": "Name of the attacking enemy",
            "example": "Goblin"
          },
          "targetId": {
            "type": "string",
            "description": "ID of the target (player characterId)",
            "example": "char-123"
          },
          "hit": {
            "type": "boolean",
            "description": "Whether the attack hit the target",
            "example": true
          },
          "attackRoll": {
            "description": "Attack roll result",
            "allOf": [
              {
                "$ref": "#/components/schemas/DiceResultDto"
              }
            ]
          },
          "damageRoll": {
            "description": "Damage roll result (only present if hit)",
            "allOf": [
              {
                "$ref": "#/components/schemas/CombatDiceResultDto"
              }
            ]
          },
          "damageTotal": {
            "type": "number",
            "description": "Total damage dealt (0 if miss)",
            "example": 5
          },
          "isCrit": {
            "type": "boolean",
            "description": "Whether the attack was a critical hit",
            "example": false
          }
        },
        "required": [
          "attackerId",
          "attackerName",
          "targetId",
          "hit"
        ]
      },
      "EndPlayerTurnResponseDto": {
        "type": "object",
        "properties": {
          "roundNumber": {
            "type": "number",
            "description": "The current round number after enemy turn",
            "example": 2
          },
          "attackLogs": {
            "description": "List of enemy attack logs in execution order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnemyAttackLogDto"
            }
          },
          "totalDamageToPlayer": {
            "type": "number",
            "description": "Total damage dealt to the player this turn",
            "example": 12
          },
          "playerDefeated": {
            "type": "boolean",
            "description": "Whether the player was defeated this turn",
            "example": false
          },
          "combatState": {
            "description": "Updated combat state after all enemy actions",
            "allOf": [
              {
                "$ref": "#/components/schemas/CombatStateDto"
              }
            ]
          }
        },
        "required": [
          "roundNumber",
          "attackLogs",
          "totalDamageToPlayer",
          "combatState"
        ]
      },
...
      "CombatEndResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation succeeded"
          },
          "message": {
            "type": "string",
            "description": "Human readable message"
          },
          "instructions": {
            "description": "Optional instructions returned after ending combat",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatEndResultDto"
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "MovementRequestDto": {
        "type": "object",
        "properties": {
          "combatantId": {
            "type": "string",
            "description": "ID of the combatant to move"
          },
          "path": {
            "description": "Path of grid positions to traverse",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridPositionDto"
            }
          }
        },
        "required": [
          "combatantId",
          "path"
        ]
      },
...
      "MovementResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether movement was successful"
          },
          "finalPosition": {
            "description": "Final position after movement",
            "allOf": [
              {
                "$ref": "#/components/schemas/GridPositionDto"
              }
            ]
          },
          "events": {
            "description": "Ordered list of events that occurred during movement",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovementEventDto"
            }
          },
          "remainingMovement": {
            "type": "number",
            "description": "Remaining movement speed after this action"
          },
          "errorMessage": {
            "type": "string",
            "description": "Error message if movement failed"
          }
        },
        "required": [
          "success",
          "finalPosition",
          "events",
          "remainingMovement"
        ]
      },
...
    }
  }
}